// src/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string | number; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /**
   * Define a Relay Cursor type:
   * https://relay.dev/graphql/connections.htm#sec-Cursor
   */
  Cursor: { input: any; output: any; }
  /** An object with an Global ID,for using in Noder interface. */
  GID: { input: any; output: any; }
  /** The builtin Time type */
  Time: { input: any; output: any; }
};
// ...
export type AppConnection = {
  __typename?: 'AppConnection';
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<AppEdge>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int']['output'];
};

/** An edge in a connection. */
export type AppEdge = {
  __typename?: 'AppEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['Cursor']['output'];
  /** The item at the end of the edge. */
  node?: Maybe<App>;
};
// ...
export type Node = {
  /** The id of the object. */
  id: Scalars['ID']['output'];
};

/** Possible directions in which to order a list of items when provided an `orderBy` argument. */
export enum OrderDirection {
  /** Specifies an ascending order for a given `orderBy` argument. */
  Asc = 'ASC',
  /** Specifies a descending order for a given `orderBy` argument. */
  Desc = 'DESC'
}

export type Org = Node & {
  __typename?: 'Org';
  apps: AppConnection;
  children?: Maybe<Array<Org>>;
  /** 系统代码 */
  code?: Maybe<Scalars['String']['output']>;
  /** 国家或地区2字码 */
  countryCode?: Maybe<Scalars['String']['output']>;
  createdAt: Scalars['Time']['output'];
  createdBy: Scalars['Int']['output'];
  deletedAt?: Maybe<Scalars['Time']['output']>;
  displaySort?: Maybe<Scalars['Int']['output']>;
  /** 默认域名 */
  domain?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  /** 是否允许解除应用策略 */
  isAllowRevokeAppPolicy: Scalars['Boolean']['output'];
  /** 分类: 根节点,组织节点 */
  kind: OrgKind;
  /** 组织名称 */
  name: Scalars['String']['output'];
  /** 管理账户 */
  owner?: Maybe<User>;
  /** 管理账户ID,如果设置则该组织将升级为根组织 */
  ownerID?: Maybe<Scalars['ID']['output']>;
  parent: Org;
  /** 父级ID,0为根组织. */
  parentID: Scalars['ID']['output'];
  /** 路径编码 */
  path?: Maybe<Scalars['String']['output']>;
  permissions: PermissionConnection;
  policies: OrgPolicyConnection;
  /** 简介 */
  profile?: Maybe<Scalars['String']['output']>;
  /** 状态 */
  status?: Maybe<OrgSimpleStatus>;
  /** 时区 */
  timezone?: Maybe<Scalars['String']['output']>;
  updatedAt?: Maybe<Scalars['Time']['output']>;
  updatedBy?: Maybe<Scalars['Int']['output']>;
  users: UserConnection;
};
// ...
export enum OrgKind {
  Org = 'org',
  Root = 'root'
}
// ...
export type OrgPolicyConnection = {
  __typename?: 'OrgPolicyConnection';
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<OrgPolicyEdge>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int']['output'];
};

/** An edge in a connection. */
export type OrgPolicyEdge = {
  __typename?: 'OrgPolicyEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['Cursor']['output'];
  /** The item at the end of the edge. */
  node?: Maybe<OrgPolicy>;
};
// ...
export enum OrgSimpleStatus {
  Active = 'active',
  Inactive = 'inactive',
  Processing = 'processing'
}
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  /** When paginating forwards, the cursor to continue. */
  endCursor?: Maybe<Scalars['Cursor']['output']>;
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars['Boolean']['output'];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars['Boolean']['output'];
  /** When paginating backwards, the cursor to continue. */
  startCursor?: Maybe<Scalars['Cursor']['output']>;
};
// ...
export type PermissionConnection = {
  __typename?: 'PermissionConnection';
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<PermissionEdge>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int']['output'];
};

/** An edge in a connection. */
export type PermissionEdge = {
  __typename?: 'PermissionEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['Cursor']['output'];
  /** The item at the end of the edge. */
  node?: Maybe<Permission>;
};
// ...
export type User = Node & {
  __typename?: 'User';
  /** 备注 */
  comments?: Maybe<Scalars['String']['output']>;
  createdAt: Scalars['Time']['output'];
  createdBy: Scalars['Int']['output'];
  /** 创建类型,邀请，注册,手工创建 */
  creationType: UserCreationType;
  deletedAt?: Maybe<Scalars['Time']['output']>;
  /** 用户设备 */
  devices?: Maybe<Array<UserDevice>>;
  /** 显示名 */
  displayName: Scalars['String']['output'];
  /** 邮箱 */
  email?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  /** 用户身份标识 */
  identities?: Maybe<Array<UserIdentity>>;
  /** 是否允许解除角色授权 */
  isAllowRevokeRole: Scalars['Boolean']['output'];
  /** 是否分配role */
  isAssignOrgRole: Scalars['Boolean']['output'];
  /** 登陆设置 */
  loginProfile?: Maybe<UserLoginProfile>;
  /** 手机 */
  mobile?: Maybe<Scalars['String']['output']>;
  permissions: PermissionConnection;
  /** 登陆名称 */
  principalName: Scalars['String']['output'];
  /** 注册时IP */
  registerIP: Scalars['String']['output'];
  /** 状态 */
  status?: Maybe<UserSimpleStatus>;
  updatedAt?: Maybe<Scalars['Time']['output']>;
  updatedBy?: Maybe<Scalars['Int']['output']>;
  /** 用户类型 */
  userType: UserUserType;
};
// ...
export type UserConnection = {
  __typename?: 'UserConnection';
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<UserEdge>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int']['output'];
};

/** UserCreationType is enum for the field creation_type */
export enum UserCreationType {
  Invitation = 'invitation',
  Manual = 'manual',
  Register = 'register'
}

export type UserDevice = Node & {
  __typename?: 'UserDevice';
  appVersion?: Maybe<Scalars['String']['output']>;
  /** 备注 */
  comments?: Maybe<Scalars['String']['output']>;
  createdAt: Scalars['Time']['output'];
  createdBy: Scalars['Int']['output'];
  deviceModel?: Maybe<Scalars['String']['output']>;
  deviceName?: Maybe<Scalars['String']['output']>;
  /** 设备唯一ID */
  deviceUID: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  /** 状态,可用或不可用及其他待确认状态 */
  status?: Maybe<UserDeviceSimpleStatus>;
  systemName?: Maybe<Scalars['String']['output']>;
  systemVersion?: Maybe<Scalars['String']['output']>;
  updatedAt?: Maybe<Scalars['Time']['output']>;
  updatedBy?: Maybe<Scalars['Int']['output']>;
  user?: Maybe<User>;
  userID?: Maybe<Scalars['ID']['output']>;
};
// ...
export type UserEdge = {
  __typename?: 'UserEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['Cursor']['output'];
  /** The item at the end of the edge. */
  node?: Maybe<User>;
};

export type UserIdentity = Node & {
  __typename?: 'UserIdentity';
  /** 用户名、邮箱、手机、unionid、qq */
  code?: Maybe<Scalars['String']['output']>;
  /** 扩展标识码,比如微信的openID */
  codeExtend?: Maybe<Scalars['String']['output']>;
  createdAt: Scalars['Time']['output'];
  createdBy: Scalars['Int']['output'];
  id: Scalars['ID']['output'];
  /** 身份标识类型 手机、邮箱、用户名、微信、qq */
  kind: UserIdentityKind;
  /** 状态,部分登陆方式需要验证通过才可启用 */
  status?: Maybe<UserIdentitySimpleStatus>;
  updatedAt?: Maybe<Scalars['Time']['output']>;
  updatedBy?: Maybe<Scalars['Int']['output']>;
  user?: Maybe<User>;
  userID?: Maybe<Scalars['ID']['output']>;
};
// ...
export type UserLoginProfile = Node & {
  __typename?: 'UserLoginProfile';
  /** 是否允许使用密码登陆控制台 */
  canLogin?: Maybe<Scalars['Boolean']['output']>;
  createdAt: Scalars['Time']['output'];
  createdBy: Scalars['Int']['output'];
  id: Scalars['ID']['output'];
  /** 最后登陆时间 */
  lastLoginAt?: Maybe<Scalars['Time']['output']>;
  lastLoginIP?: Maybe<Scalars['String']['output']>;
  /** 是否开启多因素验证 */
  mfaEnabled?: Maybe<Scalars['Boolean']['output']>;
  /** 多因素验证状态 */
  mfaStatus?: Maybe<UserLoginProfileSimpleStatus>;
  /** 下次登陆时需要重置密码 */
  passwordReset?: Maybe<Scalars['Boolean']['output']>;
  /** 设置密码:keep-保持不变,customer-客户自行设置,auto-自动生成 */
  setKind: UserLoginProfileSetKind;
  updatedAt?: Maybe<Scalars['Time']['output']>;
  updatedBy?: Maybe<Scalars['Int']['output']>;
  user?: Maybe<User>;
  userID?: Maybe<Scalars['ID']['output']>;
  /** 是否开启设备认证 */
  verifyDevice: Scalars['Boolean']['output'];
};
// ...
export enum UserSimpleStatus {
  Active = 'active',
  Inactive = 'inactive',
  Processing = 'processing'
}

/** UserUserType is enum for the field user_type */
export enum UserUserType {
  Account = 'account',
  Member = 'member'
}